refactor: use-meeting-comment-item을 widget model로 분리
