+++ src/widgets/meeting-detail/ui/meeting-comment-section/meeting-comment-item.tsx
@@ -1,20 +1,11 @@
 'use client';
 
-import { useEffect, useRef, useState } from 'react';
-
 import { Heart, MessageCircle, UserRound } from 'lucide-react';
 
-import { useAuthStore } from '@/entities/auth';
-import {
-  useCreateComment,
-  useDeleteComment,
-  useLikeComment,
-  useUpdateComment,
-} from '@/entities/comment';
-import { useModal } from '@/shared/lib/use-modal';
 import { cn } from '@/shared/lib/utils';
 import { Avatar, AvatarFallback, AvatarImage } from '@/shared/ui/avatar';
 
+import { useMeetingCommentItem } from '../../model/use-meeting-comment-item';
 import { EllipsisMenu } from '../meeting-detail-card/_components/ellipsis-menu';
 
 import { CommentDeleteModal } from './comment-delete-modal';
@@ -26,8 +17,6 @@ interface MeetingCommentItemProps {
   isReply?: boolean;
   meetingId: number;
 }
-
-// ─── MeetingCommentItem ───────────────────────────────────────
 
 export function MeetingCommentItem({
   comment,
@@ -47,93 +36,27 @@ export function MeetingCommentItem({
     replies,
   } = comment;
 
-  const [isEditing, setIsEditing] = useState(false);
-  const [editText, setEditText] = useState(content);
-  const [isReplying, setIsReplying] = useState(false);
-  const [replyText, setReplyText] = useState('');
-  const [prevServerLike, setPrevServerLike] = useState({ isLiked, likeCount });
-  const [localIsLiked, setLocalIsLiked] = useState(isLiked);
-  const [localLikeCount, setLocalLikeCount] = useState(likeCount);
-  const timerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const { isOpen: isDeleteModalOpen, open: openDeleteModal, close: closeDeleteModal } = useModal();
-
-  // 서버 데이터 변경 시 렌더 중 동기화 (useEffect 대신 React 권장 파생 상태 패턴)
-  if (prevServerLike.isLiked !== isLiked || prevServerLike.likeCount !== likeCount) {
-    setPrevServerLike({ isLiked, likeCount });
-    setLocalIsLiked(isLiked);
-    setLocalLikeCount(likeCount);
-  }
-
-  const { user } = useAuthStore();
-  const { mutate: likeComment } = useLikeComment(meetingId);
-  const { mutate: updateComment } = useUpdateComment(meetingId);
-  const { mutate: deleteComment } = useDeleteComment(meetingId);
-  const { mutate: createComment } = useCreateComment(meetingId, {
-    nickname: user?.name ?? '',
-    profileUrl: user?.image ?? null,
-  });
+  const {
+    isEditing,
+    setIsEditing,
+    editText,
+    setEditText,
+    isReplying,
+    setIsReplying,
+    replyText,
+    setReplyText,
+    localIsLiked,
+    localLikeCount,
+    isDeleteModalOpen,
+    openDeleteModal,
+    closeDeleteModal,
+    handleLike,
+    handleEditSubmit,
+    handleDelete,
+    handleReplySubmit,
+  } = useMeetingCommentItem({ commentId: id, meetingId, content, isLiked, likeCount });
 
   const isPending = id < 0;
-
-  useEffect(() => {
-    return () => {
-      if (timerRef.current) clearTimeout(timerRef.current);
-    };
-  }, []);
-
-  const handleLike = () => {
-    const nextIsLiked = !localIsLiked;
-    setLocalIsLiked(nextIsLiked);
-    setLocalLikeCount((prev) => (nextIsLiked ? prev + 1 : prev - 1));
-
-    if (timerRef.current) clearTimeout(timerRef.current);
-    timerRef.current = setTimeout(() => {
-      if (nextIsLiked === prevServerLike.isLiked) return;
-
-      likeComment(
-        { commentId: id, isLiked: localIsLiked },
-        {
-          onError: () => {
-            setLocalIsLiked(isLiked);
-            setLocalLikeCount(likeCount);
-          },
-        }
-      );
-    }, 300);
-  };
-
-  const handleEditSubmit = () => {
-    if (!editText.trim() || editText === content) {
-      setIsEditing(false);
-      return;
-    }
-    updateComment(
-      { commentId: id, payload: { content: editText.trim() } },
-      { onSuccess: () => setIsEditing(false) }
-    );
-  };
-
-  const handleDelete = () => {
-    deleteComment(id, {
-      onSuccess: closeDeleteModal,
-    });
-  };
-
-  const handleReplySubmit = () => {
-    if (!replyText.trim()) return;
-    const savedText = replyText;
-    setIsReplying(false);
-    setReplyText('');
-    createComment(
-      { content: savedText, parentId: id },
-      {
-        onError: () => {
-          setIsReplying(true);
-          setReplyText(savedText);
-        },
-      }
-    );
-  };
 
   return (
     <div>
